requirements.txt
fastapi
uvicorn
groq
python-dotenv
faiss-cpu
sentence-transformers
pyvi
rank-bm25
PyMuPDF
python-multipart
Pillow
typing-extensions
tensorflow
